refactor: 메뉴 가격 정책 리팩터링
